+++ examples/cpu/cpu_allreduce_bf16_test.cpp
@@ -28,15 +28,15 @@
 #define CHECK_ERROR(send_buf, recv_buf, comm) \
     { \
         /* https://www.mcs.anl.gov/papers/P4093-0713_1.pdf */ \
-        size_t comm_size = comm.size(); \
+        int comm_size = comm.size(); \
         double log_base2 = log(comm_size) / log(2); \
         double g = (log_base2 * BF16_PRECISION) / (1 - (log_base2 * BF16_PRECISION)); \
         for (size_t i = 0; i < COUNT; i++) { \
-            double expected = ((comm_size * (comm_size - 1) / 2) + ((float)(i) * comm_size)); \
+            double expected = ((comm_size * (comm_size - 1) / 2) + ((float)(i)*comm_size)); \
             double max_error = g * expected; \
             if (fabs(max_error) < fabs(expected - recv_buf[i])) { \
                 printf( \
-                    "[%zu] got recv_buf[%zu] = %0.7f, but expected = %0.7f, max_error = %0.16f\n", \
+                    "[%d] got recv_buf[%zu] = %0.7f, but expected = %0.7f, max_error = %0.16f\n", \
                     comm.rank(), \
                     i, \
                     recv_buf[i], \
@@ -50,7 +50,6 @@
 using namespace std;
 
 int main() {
-
     const size_t count = 4096;
 
     size_t idx = 0;
@@ -93,12 +92,9 @@ int main() {
     else {
         cout << "BF16 is enabled\n";
         convert_fp32_to_bf16_arrays(send_buf, send_buf_bf16, count);
-        ccl::allreduce(send_buf_bf16,
-                       recv_buf_bf16,
-                       count,
-                       ccl::datatype::bfloat16,
-                       ccl::reduction::sum,
-                       comm).wait();
+        ccl::allreduce(
+            send_buf_bf16, recv_buf_bf16, count, ccl::datatype::bfloat16, ccl::reduction::sum, comm)
+            .wait();
         convert_bf16_to_fp32_arrays(recv_buf_bf16, recv_buf, count);
         CHECK_ERROR(send_buf, recv_buf, comm);
